+++ doc/source/changes.rst
@@ -1,10 +1,13 @@
 Change Log
 ==========
 
-0.8.4 - Future release
+0.9.0 - Future release
 ----------------------
 * Replace examples documentation with more in depth usage guide (:doc:`usage`).
 * Add documentation for performing a release of the library (:doc:`release`).
+* Move enabling/disabling certificate verification to the :class:`warthog.client.WarthogClient`
+  class so that using self-signed certificates is less of a hassle and requires
+  less code.
 
 0.8.3 - 2015-03-18
 ------------------